extracted validItem and invalidItem into their own file and import paths
