+++ backend/src/models/Item/index.ts
@@ -1,2 +1,3 @@
 export { default } from "./item-model";
 export { ItemI } from "./item-types";
+export { validItem, invalidItem } from "./item-examples";
