+++ backend/src/models/Item/item-examples.ts
@@ -1,0 +1,10 @@
+export const validItem = {
+  name: "Honda Civic",
+  description: "We want a new car!",
+  price: 20000,
+};
+
+export const invalidItem = {
+  name: "Honda Civic",
+  price: -1,
+};
