+++ backend/src/models/Item/item-types.ts
@@ -13,14 +13,3 @@ export interface ItemI extends Document {
   reservedOn?: Date;
   purchasedOn?: Date;
 }
-
-export const validItem = {
-  name: "Honda Civic",
-  description: "We want a new car!",
-  price: 20000,
-};
-
-export const invalidItem = {
-  name: "Honda Civic",
-  price: -1,
-};
